kb_mcp/requirements.txt
mcp>=0.1.0
qdrant-client>=1.7.0
python-arango>=7.9.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyyaml>=6.0
